requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
async-generator==1.10
async-timeout==3.0.1
asyncio-extras==1.3.2
asyncpraw==7.5.0
asyncprawcore==2.3.0
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.11
discord.py==1.7.3
dnspython==2.2.0
idna==3.3
install==1.3.5
multidict==6.0.2
praw==7.5.0
prawcore==2.3.0
pycodestyle==2.8.0
pymongo==4.0.1
python-dotenv==0.19.2
requests==2.27.1
toml==0.10.2
typing_extensions==4.0.1
update-checker==0.18.0
urllib3==1.26.8
websocket-client==1.2.3
yarl==1.7.2
